backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pandas==2.2.3
anthropic==0.43.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
aiofiles==24.1.0
pytest==7.4.4
typing_extensions>=4.12.0
